backend/pyproject.toml
[tool.poetry]
name = "today-lunch"
version = "0.1.0"
description = ""
authors = ["byungyong <quddyddlrk@naver.com>"]
readme = "README.md"
packages = [{include = "today_lunch"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
django-cors-headers = "^4.1.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
